backend_ai/pyproject.toml
[project]
name = "backend-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.9",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "beanie==1.30.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "dlib==20.0.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "face-recognition==1.3.0",
    "face-recognition-models==0.3.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "google-auth==2.40.3",
    "google-genai==1.20.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "lazy-model==0.2.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "motor==3.7.1",
    "numpy==2.3.0",
    "pillow==11.2.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pygments==2.19.1",
    "pymongo==4.13.2",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "rich==14.0.0",
    "rich-toolkit==0.14.1",
    "rsa==4.9.1",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "wheel==0.45.1",
]
